+++ filen-cli/CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-- /
+### Added
+
+- rclone commands (`mount`, `serve`) are more customizable via `--cache-size` and `--transfers` flags
+  as well as the option to take any number of custom Rclone options
 
 ## 0.2.4 - 2026-01-20